% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fit_functions.R
\name{adam}
\alias{adam}
\title{ADAM optimizer.}
\usage{
adam(theta0, fn, gr, nsteps, alpha = 0.001, beta = c(0.9, 0.999), eps = 1e-08)
}
\arguments{
\item{theta0}{Initial parameter value (numeric vector).}

\item{fn}{Objective function to minimize.}

\item{gr}{Gradient function of \code{fn}.}

\item{nsteps}{Number of optimization steps to perform.}

\item{alpha, beta, eps}{Algorithm tuning parameters.  \code{alpha} and \code{eps} are scalars, \code{beta} is a numeric vector of length 2.}
}
\value{
A list with elements:
\describe{
\item{\code{par}}{The parameter value at the minimum of the \code{nsteps} optimization steps.}
\item{\code{objective}}{The minimum value of the objective function.}
\item{\code{Theta}}{A matrix where each row is the parameter value at the given step.}
\item{\code{Fn}}{A vector where each element is the objective function value at the given step.}
}
}
\description{
ADAM optimizer.
}
