% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_factor.R
\name{bin_factor}
\alias{bin_factor}
\title{Calculate the bin size correction factor for the \code{LP} estimator.}
\usage{
bin_factor(B)
}
\arguments{
\item{B}{Bin size (integer).}
}
\value{
The bin size correction factor \code{log(B) - digamma(B)}.
}
\description{
Calculate the bin size correction factor for the \code{LP} estimator.
}
