% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{binning}
\usage{
binning(x, binSize)
}
\arguments{
\item{x}{numeric \code{vector} to process}

\item{binSize}{length of each bin}
}
\value{
a \code{matrix} or \code{vector} with average values per bin
}
\description{
Compute average values of a series in pre-determined bins (column-wise subsets).
The bin size can be determined either directly or by specifying the number of bins.
}
