% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_basis.R
\name{fft_basis}
\alias{fft_basis}
\title{Calculate the frequency basis of the discrete Fourier transform.}
\usage{
fft_basis(fs, N, frng)
}
\arguments{
\item{fs}{Sampling frequency.}

\item{N}{Number of time domain observations.}

\item{frng}{Optional frequency range (vector of length 2) on which to restrict the basis.}
}
\value{
The one-sided frequency basis \code{freq = 1:floor(N/2)/N * fs}, optionally restricted to \code{frng}.
}
\description{
Calculate the frequency basis of the discrete Fourier transform.
}
