% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_fseq.R
\name{fft_fseq}
\alias{fft_fseq}
\title{Calculate the frequency basis of the Fourier transform.}
\usage{
fft_fseq(fs, N, frng)
}
\arguments{
\item{fs}{Sampling frequency.}

\item{N}{Number of time domain observations.}

\item{frng}{Optional frequency range (vector of length 2) on which to restrict the basis.}
}
\value{
The one-sided frequency basis \code{fseq = 1:floor(N/2)/N * fs}, optionally restricted to \code{frng}.
}
\description{
Calculate the frequency basis of the Fourier transform.
}
