% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_sin.R
\name{fft_sin}
\alias{fft_sin}
\title{FFT (discrete Fourier transform) of sine wave noise}
\usage{
fft_sin(fseq, N, f0, Q, fs, unit_conversion)
}
\arguments{
\item{fseq}{A sequence of frequency points (must be points in the frequency basis)}

\item{N}{total length of the frequency domain}

\item{f0}{Resonance frequency}

\item{Q}{Quality factor}

\item{fs}{Sampling frequency}

\item{unit_conversion}{TRUE/FALSE, if TRUE, use fm2/Hz}
}
\description{
FFT (discrete Fourier transform) of sine wave noise
}
