% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherG.R
\name{fisherG}
\alias{fisherG}
\title{Calculate the tail probability of Fisher's G statistic.}
\usage{
fisherG(a, q, log_sort = TRUE)
}
\arguments{
\item{a}{Probability threshold (scalar between 0 and 1).}

\item{q}{Number of variables from which Fisher's G is computed (see 'Details').}

\item{log_sort}{If \code{TRUE}, reorders the terms of the alternating series to minimize numerical overflow.}
}
\value{
Tail probability of Fisher's G statistic: \code{Pr(G > a)}.  Attempts to fail with \code{NA} when numerical underflow (\code{prob < 0}) is detected.
}
\description{
Calculate the tail probability of Fisher's G statistic.
}
\details{
Let \verb{U = U_0 < ... < U_q}, where \verb{U_1, ... ,U_\{q-1\}} are the \emph{order statistics} of \code{n-1} iid Uniforms, \code{U_0 = 0}, and \code{U_q = 1}.  Then Fisher's G statistic is defined as \code{G = max(diff(U))}.

It turns out that if \code{Yf} are \code{n} periodogram ordinates corresponding to white noise, then \code{G = max(Yf)/sum(Yf)}, such that Fisher's G statistic can be used to test for hidden periodic components in \code{Yf}.
}
