% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSHOW.R
\name{fitSHOW}
\alias{fitSHOW}
\title{fit simulated datasets to get fitted parameters}
\usage{
fitSHOW(f, rfreq, fs, f0, Q, k, T, Aw, binSize = 100,
  method = c("LP_nlp", "MLE_nlp", "NLS_nlp"))
}
\arguments{
\item{f}{Vector of frequencies, usually from f0 - f0/sqrt(2) to f0 + f0/sqrt(2).}

\item{rfreq}{Vector of exponential random variables Exp(1) with the same length as f.}

\item{fs}{Sampling frequency, Hz.}

\item{f0}{Resonance frequency, Hz.}

\item{Q}{Quality factor.}

\item{k}{Cantilever stiffness, N/m.}

\item{T}{Temperature, Kelvin.}

\item{Aw}{White noise psd.}

\item{binSize}{Integer number, bin size.}

\item{method}{Fitting method, i.e. LP_nlp, MLE_nlp, NLS_nlp.}

\item{Kb}{Boltzmann's constant.}
}
\description{
fit simulated datasets to get fitted parameters
}
