% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fit_functions.R
\name{fn_fixed}
\alias{fn_fixed}
\title{Evaluate a \pkg{TMB} objective function at fixed parameter values.}
\usage{
fn_fixed(phi, obj, fixed, phi0)
}
\description{
Evaluate a \pkg{TMB} objective function at fixed parameter values.
}
