% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_freqNoise_fromAPSD.R
\name{generate_freqNoise_fromAPSD}
\alias{generate_freqNoise_fromAPSD}
\title{Generate_FreqNoise_FromAPSD}
\usage{
generate_freqNoise_fromAPSD(xAPSD, yAPSD, SF, T, Q, CORRUPT, displ,
  varargin)
}
\arguments{
\item{xAPSD}{Frequency vector of yAPSD (Hz) : [1 2 3 ... N/2]*FR this vector need NOT be linearly spaced.}

\item{yAPSD}{Average power spectral density (u^2/Hz).}

\item{SF}{Sampling frequency (Hz).}

\item{T}{Total duration (s).}

\item{Q}{Quality factor to scale sine-wave amplitude by.}

\item{displ}{1: display information, 0, don't display information.}
}
\value{
A list of xTime = time (s) and yTime = amplitude (u)
}
\description{
Generate_FreqNoise_FromAPSD
}
