% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fit.R
\name{get_par}
\alias{get_par}
\title{Recover the SHOW parameters from \code{theta}.}
\usage{
get_par(theta, Temp)
}
\arguments{
\item{theta}{Parameter vector as defined by \code{get_theta}.}

\item{Temp}{Temperature (Kelvin).}
}
\value{
Numeric vector with named elements \verb{(k, f0, Q, Sw)}.
}
\description{
Recover the SHOW parameters from \code{theta}.
}
