% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fit_functions.R
\name{get_par_shof}
\alias{get_par_shof}
\title{Recover the SHOF parameters from \code{theta}.}
\usage{
get_par_shof(theta, Temp)
}
\arguments{
\item{theta}{Parameter vector (f0, Q, Rf, alpha, tau).}

\item{Temp}{Temperature (Kelvin).}
}
\value{
Numeric vector with named elements \verb{(k, f0, Q, Af, alpha)}.
}
\description{
Recover the SHOF parameters from \code{theta}.
}
