% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fit_functions.R
\name{get_par_showf}
\alias{get_par_showf}
\title{Recover the SHOWF parameters from \code{theta}.}
\usage{
get_par_showf(theta, Temp)
}
\arguments{
\item{theta}{Parameter vector (f0, Q, Rw, Rf, alpha, tau).}

\item{Temp}{Temperature (Kelvin).}
}
\value{
Numeric vector with named elements \verb{(k, f0, Q, Sw)}.
}
\description{
Recover the SHOWF parameters from \code{theta}.
}
