% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fit.R
\name{get_theta}
\alias{get_theta}
\title{Get \code{theta} from SHOW parameters.}
\usage{
get_theta(k, f0, Q, Sw, Temp)
}
\arguments{
\item{k}{Cantilever stiffness (N/m).}

\item{f0}{Resonance frequency (Hz).}

\item{Q}{Quality factor (unitless).}

\item{Sw}{White noise power (m^2/Hz).}

\item{Temp}{Temperature (Kelvin).}
}
\value{
Numeric vector with named elements \verb{(f0, Q, Rw, tau)}.
}
\description{
The parameter vector \code{theta} is defined as\preformatted{theta = (f0, Q, Rw = Sw/tau, tau),
}

where \verb{tau = KbT/(pi k f0 Q)}.
}
