% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSHO.R
\name{getSHO}
\alias{getSHO}
\title{Generates an instance of the SHO function as well as a math function funcSHO.}
\usage{
getSHO(f, f0, Q, k)
}
\value{
...
}
\description{
Generates an instance of the SHO function as well as a math function funcSHO.
}
