% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fit_functions.R
\name{gr_fixed}
\alias{gr_fixed}
\title{Evaluate a \pkg{TMB} gradient function at fixed parameter values.}
\usage{
gr_fixed(phi, obj, fixed, phi0)
}
\description{
Evaluate a \pkg{TMB} gradient function at fixed parameter values.
}
