% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logSHOW.R
\name{logSHOW}
\alias{logSHOW}
\title{log Power spectral density of continuous-time SHO model.}
\usage{
logSHOW(fseq, f0, Q, tau, Rw)
}
\arguments{
\item{fseq}{Frequency or a vector of frequencies in Hz.}

\item{f0}{Resonance frequency f0.}

\item{Q}{Quality factor.}

\item{tau}{kB * Temp / (k*pi*f0*Q), tau should supplied with unit_conversion if it is required}

\item{Rw}{Aw/tau}
}
\description{
log Power spectral density of continuous-time SHO model.
}
