% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_psd_model.R
\name{make_psd_model}
\alias{make_psd_model}
\title{Create the \pkg{TMB} wrapper file for parameter estimation with user-defined models.}
\usage{
make_psd_model(model, header, class, ctor, standalone = TRUE, path = NULL)
}
\arguments{
\item{model}{Name of model.}

\item{header}{Name of model header file including extension.  If missing defaults to \verb{\{model\}.hpp}.  (See 'Details').}

\item{class}{Name of class definition, with enclosing namespace if it exists.  If missing defaults to \code{{model}}.  (See 'Details').}

\item{ctor}{Name of external class constructor, with enclosing namespace if it exists.  If missing defaults to \verb{make_\{model\}}.  (See 'Details').}

\item{standalone}{If \code{TRUE} creates a standalone \code{cpp} file to pass to \code{\link[TMB:compile]{TMB::compile()}}.  Otherwise, creates an \code{hpp} header file to be placed in a package created with \pkg{TMBtools}. (See 'Details').}
}
\value{
Vector of strings each containing one line of the \pkg{TMB} wrapper file.
}
\description{
Create the \pkg{TMB} wrapper file for parameter estimation with user-defined models.
}
