% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{noise_sin}
\alias{noise_sin}
\title{noise_sin: This function returns one-sided, unfolded, FFT of sine, with DC component removed.}
\usage{
noise_sin(f_noise, A_noise, T, SF, detrend_flag)
}
\arguments{
\item{T}{Total time.}

\item{SF}{Sampling Freq (Hz).}

\item{detrend_flag}{1 = True, 0 = False.}

\item{A}{Amplitude.}

\item{f}{Frequency.}
}
\value{
A list of 
xTime: Time values, 
yTime: Sine signal values, 
xFreq: Frequencies,
yFreq: First freq. up to Nyquist freq of: fft(yTime)/(sqrt(FR)*N)
}
\description{
noise_sin: This function returns one-sided, unfolded, FFT of sine, with DC component removed.
}
\details{
NOTE: Removes DC component, select first frequency up to the Nyquist frequency. NOTE: Uses even PSD's only so it will round yTime to multiple of 2
}
