% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ou_sim.R
\name{ou_sim}
\alias{ou_sim}
\title{Generate observations from an Ornstein-Uhlenbeck process at regular time intervals.}
\usage{
ou_sim(gamma, mu, sigma, dt, n_obs, X0)
}
\arguments{
\item{gamma}{Scalar mean reversion parameter (see \strong{Details}).}

\item{mu}{Scalar mean parameter (see \strong{Details}).}

\item{sigma}{Scalar diffusion parameter (see \strong{Details}).}

\item{dt}{Interobservation time.}

\item{n_obs}{Number of observations to generate.}

\item{X0}{Initial value of the process at time \code{t = 0}.  If missing sampled from the OU stationary distribution.}
}
\value{
A vector of \code{n_obs} observations of the process at times \verb{t = dt, 2*dt, ..., n_obs * dt}.
}
\description{
Generate observations from an Ornstein-Uhlenbeck process at regular time intervals.
}
\details{
The Ornstein-Uhlenbeck (OU) process satisfies a stochastic differential equation of the form\preformatted{dX_t = -gamma * (X_t - mu) dt + sigma dB_t.
}

It is a stationary Gaussian Markov process with transition density\preformatted{X_s+t | X_s ~ N( rho_t * (X_s - mu) + mu, tau^2 * (1-rho_t^2) ),
}

where \code{rho_t = exp(-gamma * t)} and \code{tau^2 = sigma^2/(2*gamma)}.  Its stationary distribution is \code{X_t ~ N(mu, tau^2)}.
}
