% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Calculate the one-sided periodogram of a time series.}
\usage{
periodogram(Xt, fs)
}
\arguments{
\item{Xt}{Vector of time series observations.}

\item{fs}{Sampling frequecy (Hz).}
}
\value{
Data frame with columns \code{freq} and \code{Ypsd}.
}
\description{
Calculate the one-sided periodogram of a time series.
}
\details{
For \code{length(Xt) = N = 2*K}, we have\preformatted{freq = 1:(K-1)/N * fs,     Ypsd = 1/(N*fs) * abs(fft(Xt)[2:K])^2.
}
}
