% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram_fft.R
\name{periodogram_fft}
\alias{periodogram_fft}
\title{Calculates the one-sided periodogram from given FFT.}
\usage{
periodogram_fft(yFFT, SF_s, T_s)
}
\arguments{
\item{yFFT}{FFT of the original time series.}

\item{SF_s}{Sampling frequecy (Hz).}

\item{T_s}{Total time (seconds).}
}
\value{
List of xFreq: One-sided periodogram frequencies and yFreq: One-sidede periodogram ordinates.
}
\description{
Calculates the one-sided periodogram from given FFT.
}
