% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_denoise.R
\name{psd_denoise}
\alias{psd_denoise}
\title{Remove periodic noise from the periodogram.}
\usage{
psd_denoise(Ypsd, psd, alpha = 0.01, method = c("fisherG", "BH"))
}
\arguments{
\item{Ypsd}{Vector of periodogram ordinates.}

\item{psd}{PSD at the values in \code{Ypsd}, i.e., theoretical expectation of \code{Ypsd} in the absence of noise.}

\item{alpha}{Type-I error tolerance level.}

\item{method}{Denoising method to use.  Either \code{fisherG} for Fisher's G-statistic test, or \code{BH} for the Benjamini-Hochberg method, in which case \code{alpha} specifies the false discovery rate.}
}
\value{
The vector \code{Ypsd} with outliers replaced by random Exponential draws with mean given by \code{psd}.
}
\description{
Remove periodic noise from the periodogram.
}
