% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_model.R
\name{psd_model}
\alias{psd_model}
\title{Base class for PSD models.}
\description{
Base class for PSD models.

Base class for PSD models.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{est_type}}{Estimator type: "lp", "nls", or "mle".}

\item{\code{bin_size}}{Bin size.}

\item{\code{bin_type}}{Bin type: "mean" or "median".}

\item{\code{freq}}{Frequency vector.}

\item{\code{Ypsd}}{Periodogram vector.}

\item{\code{fbar}}{Binned frequency vector.}

\item{\code{Ybar}}{Binned periodogram vector.}

\item{\code{Zbar}}{Log binned periodogram vector.}

\item{\code{n_phi}}{Dimension of \code{phi}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_est}{\code{psd_model$set_est()}}
\item \href{#method-set_psd}{\code{psd_model$set_psd()}}
\item \href{#method-nlp}{\code{psd_model$nlp()}}
\item \href{#method-nll}{\code{psd_model$nll()}}
\item \href{#method-resid}{\code{psd_model$resid()}}
\item \href{#method-ufun}{\code{psd_model$ufun()}}
\item \href{#method-zeta}{\code{psd_model$zeta()}}
\item \href{#method-vcov}{\code{psd_model$vcov()}}
\item \href{#method-new}{\code{psd_model$new()}}
\item \href{#method-to_est}{\code{psd_model$to_est()}}
\item \href{#method-clone}{\code{psd_model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_est"></a>}}
\if{latex}{\out{\hypertarget{method-set_est}{}}}
\subsection{Method \code{set_est()}}{
Specify the estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$set_est(
  est_type = c("lp", "nls", "mle"),
  bin_size = 100,
  bin_type = c("mean", "median")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{est_type}}{Type of estimator: "lp", "nls", or "mle".}

\item{\code{bin_size}}{Bin size.}

\item{\code{bin_type}}{Type of bin: "mean" or "median".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_psd"></a>}}
\if{latex}{\out{\hypertarget{method-set_psd}{}}}
\subsection{Method \code{set_psd()}}{
Set the periodogram data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$set_psd(freq, Ypsd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{freq}}{Frequency vector.}

\item{\code{Ypsd}}{Periodogram vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nlp"></a>}}
\if{latex}{\out{\hypertarget{method-nlp}{}}}
\subsection{Method \code{nlp()}}{
Negative profile loglikelihood constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$nlp(phi0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi0}}{Vector with which to initialize \code{phi}.  If missing, a vector of zeros.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} object.  Relevant methods are: \code{nlp$fn(phi)}, \code{nlp$gr(phi)}, \code{nlp$he(phi)}, and \code{nlp$simulate(phi)$zeta}, which returns the conditional optimum of \code{zeta} given \code{phi}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nll"></a>}}
\if{latex}{\out{\hypertarget{method-nll}{}}}
\subsection{Method \code{nll()}}{
Negative loglikelihood constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$nll(phi0, zeta0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi0}}{Vector with which to initialize \code{phi}.  If missing, a vector of zeros.}

\item{\code{zeta0}}{Scalar with which to initialize \code{zeta}.  If missing, zero.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} object.  Relevant methods are: \code{nll$fn(phi)}, \code{nll$gr()}, and \code{nll$he()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resid"></a>}}
\if{latex}{\out{\hypertarget{method-resid}{}}}
\subsection{Method \code{resid()}}{
Residual constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$resid(phi0, zeta0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi0}}{Vector with which to initialize \code{phi}.  If missing, a vector of zeros.}

\item{\code{zeta0}}{Scalar with which to initialize \code{zeta}.  If missing, zero.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} object.  Relevant methods are: \code{resid$fn(phi)} and \code{resid$gr(phi)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ufun"></a>}}
\if{latex}{\out{\hypertarget{method-ufun}{}}}
\subsection{Method \code{ufun()}}{
Normalized PSD constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$ufun(phi0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi0}}{Vector with which to initialize \code{phi}.  If missing, a vector of zeros.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} object.  Relevant methods are: \code{ufun$fn(phi)} and \code{ufun$gr(phi)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-zeta"></a>}}
\if{latex}{\out{\hypertarget{method-zeta}{}}}
\subsection{Method \code{zeta()}}{
Calculate the optimal value of the log normalizing constant.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$zeta(phi0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi0}}{Vector with which to initialize \code{phi}.  If missing, a vector of zeros.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} object.  Relevant methods are: \code{zeta$fn(phi)}, \code{zeta$gr()}, and \code{zeta$he()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vcov"></a>}}
\if{latex}{\out{\hypertarget{method-vcov}{}}}
\subsection{Method \code{vcov()}}{
Calculate the variance estimator for the parameter fitting method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$vcov(phi, zeta, to_theta = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi}}{Vector of normalized PSD parameters.}

\item{\code{zeta}}{Log of normalization parameter.}

\item{\code{to_theta}}{If \code{TRUE}, calculates the variance in the inferential basis.  Otherwise, uses the computational basis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A variance matrix of size \verb{n_phi x n_phi}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Model object constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$new(
  freq,
  Ypsd,
  est_type = c("lp", "nls", "mle"),
  bin_size = 100,
  bin_type = c("mean", "median"),
  ctor_args = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{freq}}{Frequency vector.}

\item{\code{Ypsd}}{Periodogram vector.}

\item{\code{est_type}}{Type of estimator: "lp", "nls", or "mle".}

\item{\code{bin_size}}{Bin size.}

\item{\code{bin_type}}{Type of bin: "mean" or "median".}

\item{\code{ctor_args}}{Optional list of additional arguments to TMB \code{UFun()} constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_est"></a>}}
\if{latex}{\out{\hypertarget{method-to_est}{}}}
\subsection{Method \code{to_est()}}{
Method to convert the fitted parameter \code{phi} to estimates coefficients and covariances (either on the natural scale or on the log scale) for report.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$to_est(phi, to_theta = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi}}{Value of \code{phi}.}

\item{\code{to_theta}}{If \code{TRUE}, calculates the variance in the inferential basis.  Otherwise, uses the computational basis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with elements \code{coef} and \code{vcov}. If \code{to_theta = TRUE}, in the inferential basis, otherwise in the computational basis.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psd_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
