% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_sho.R
\name{psd_sho}
\alias{psd_sho}
\title{Power spectral density of SHO model.}
\usage{
psd_sho(f, f_0, Q, k, kb, T)
}
\arguments{
\item{f}{Frequency or a vector of frequencies in Hz.}

\item{f_0}{Resonance frequency f_0.}

\item{Q}{Quality factor.}

\item{k}{Cantilever stiffness.}

\item{kb}{Boltzmann constant.}

\item{T}{Temperature in Kelvin.}
}
\description{
Power spectral density of SHO model.
}
