% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realPSD-package.R
\docType{package}
\name{realPSD-package}
\alias{realPSD}
\alias{realPSD-package}
\title{realPSD: What the package does (short line)}
\description{
More about what it does (maybe more than one line).
}
\author{
\strong{Maintainer}: Martin Lysy \email{f28zhu@uwaterloo.ca}

Authors:
\itemize{
  \item Feiyu Zhu
}

}
\keyword{internal}
