% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shof_fit_lp.R
\name{shof_fit_lp}
\alias{shof_fit_lp}
\title{SHOF fit method log-periodogram (LP). Fit SHOWF via LP method.}
\usage{
shof_fit_lp(
  fseq,
  Ypsd,
  fs,
  Temp,
  bin_size,
  bin_type,
  phi0,
  Sw0,
  tau0,
  fit_type = c("direct", "incremental"),
  optimizer = c("optim", "Adam"),
  vcov = FALSE,
  ...
)
}
\arguments{
\item{fseq}{Frequency vector (units: Hz).}

\item{Ypsd}{PSD vector.}

\item{fs}{Sampling frequency.}

\item{Temp}{True temperature.}

\item{bin_size}{Bin size.}

\item{bin_type}{Either "mean" or "median".  The former is more efficient, the latter is more robust, so better for preliminary SHOW fit to denoise against.}

\item{phi0}{Parameter value (transformed scale) to initialize optimization.}

\item{fit_type}{If "direct", fit all parameters at once.  If "incremental", do one, then two, then three, etc.}

\item{optimizer}{Either "optim" (in R) or "Adam" (supplied by realPSD). For now, Adam's hyperparameter tunning is not fully supported. Learning rate and nsteps are preset. \code{Adam} is only used for \code{direct} fitting.}

\item{vcov}{If TRUE, return the numerical Hessian matrix of the original SHOW parameter k, f0, Q, Af}

\item{...}{Additional arguments to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
A list with elements
\describe{
\item{\code{par}}{The fitted parameter value which is the minimizer of the objective function.}
\item{\code{Temp}}{The temperature constant used in calculation, Kelvin.}
\item{\code{value}}{The value of the objective function (negative loglikelihood) at the given \code{par}.}
\item{\code{cov}}{The numerical variance-covariance matrix for all the parameters (original scale).}
\item{\code{exitflag}}{The exit flag given by the optimizer, for debug purposes}
}
}
\description{
SHOF fit method log-periodogram (LP). Fit SHOWF via LP method.
}
