% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fit.R
\name{show_fit}
\alias{show_fit}
\title{Fit SHOW model}
\usage{
show_fit(
  fseq,
  Ypsd,
  fs,
  Temp,
  method = c("lp", "nls", "mle"),
  bin_size,
  bin_type,
  phi0,
  fit_type = c("direct", "incremental"),
  optimizer = c("optim", "Adam"),
  vcov = FALSE,
  get_jac = FALSE,
  ...
)
}
\arguments{
\item{fseq}{Frequency vector (units: Hz).}

\item{Ypsd}{PSD vector, returned by realPSD::periodogram() which is |fft(X_t)|^2/(N*fs).}

\item{fs}{Sampling frequency.}

\item{Temp}{True temperature.}

\item{method}{One of the three fitting methods "LP", "NLS" and "MLE".}

\item{bin_size}{Bin size.}

\item{bin_type}{Either "mean" or "median".  The former is more efficient, the latter is more robust, so better for preliminary SHOW fit to denoise against.}

\item{phi0}{Parameter value (transformed scale) to initialize optimization.}

\item{fit_type}{If "direct", fit all parameters at once.  If "incremental", do one, then two, then three, etc.}

\item{optimizer}{Either "optim" (in R) or "Adam" (supplied by realPSD). For now, Adam's hyperparameter tunning is not fully supported. Learning rate and nsteps are preset. \code{Adam} is only used for \code{direct} fitting.}

\item{vcov}{If TRUE, return the numerical variance-covariance matrix for all the parameters (original scale)}

\item{get_jac}{If TRUE, return the numerical Jacobian of the SHOW parameters k, f0, Q, only valid if method == "NLS".}

\item{...}{Additional arguments to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
A list with elements
\describe{
\item{\code{par}}{The fitted parameter value which is the minimizer of the objective function.}
\item{\code{Temp}}{The temperature constant used in calculation, Kelvin.}
\item{\code{value}}{The value of the objective function (negative loglikelihood) at the given \code{par}.}
\item{\code{cov}}{The numerical variance-covariance matrix for all the parameters (original scale).}
\item{\code{jac}}{The numerical Jacobian matrix of the SHOW model parameter f0, Q, Rw, only vailable if method == "NLS".}
\item{\code{exitflag}}{The exit flag given by the optimizer, for debug purposes}
}
}
\description{
Fit SHOW model
}
\note{
Ypsd used as input argument should be returned by realPSD::periodogram which is 1/(N*fs) * |fft(X_t)|^2, but TMB method to fit the SHOW model assumes unscaled periodogram, so we first transform Ypsd in the function. The TMB fitting methods should be modified later to account for this.
}
