% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_model.R
\name{show_model}
\alias{show_model}
\title{SHOW model class.}
\description{
SHOW model class.

SHOW model class.
}
\section{Super class}{
\code{\link[realPSD:psd_model]{realPSD::psd_model}} -> \code{show_model}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Temp}}{Temperature of the system.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{show_model$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="realPSD" data-topic="psd_model" data-id="initialize">}\href{../../realPSD/html/psd_model.html#method-initialize}{\code{realPSD::psd_model$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="realPSD" data-topic="psd_model" data-id="nll">}\href{../../realPSD/html/psd_model.html#method-nll}{\code{realPSD::psd_model$nll()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="realPSD" data-topic="psd_model" data-id="nlp">}\href{../../realPSD/html/psd_model.html#method-nlp}{\code{realPSD::psd_model$nlp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="realPSD" data-topic="psd_model" data-id="resid">}\href{../../realPSD/html/psd_model.html#method-resid}{\code{realPSD::psd_model$resid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="realPSD" data-topic="psd_model" data-id="set_est">}\href{../../realPSD/html/psd_model.html#method-set_est}{\code{realPSD::psd_model$set_est()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="realPSD" data-topic="psd_model" data-id="set_psd">}\href{../../realPSD/html/psd_model.html#method-set_psd}{\code{realPSD::psd_model$set_psd()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="realPSD" data-topic="psd_model" data-id="ufun">}\href{../../realPSD/html/psd_model.html#method-ufun}{\code{realPSD::psd_model$ufun()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="realPSD" data-topic="psd_model" data-id="vcov">}\href{../../realPSD/html/psd_model.html#method-vcov}{\code{realPSD::psd_model$vcov()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="realPSD" data-topic="psd_model" data-id="zeta">}\href{../../realPSD/html/psd_model.html#method-zeta}{\code{realPSD::psd_model$zeta()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{show_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
