% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fit.R
\name{show_psd}
\alias{show_psd}
\title{Calculate SHOW PSD.}
\usage{
show_psd(fseq, k, f0, Q, Sw, Temp)
}
\arguments{
\item{fseq}{Frequency vector (units: Hz).}

\item{k}{Cantilever stiffness (N/m).}

\item{f0}{Resonance frequency (Hz).}

\item{Q}{Quality factor (unitless).}

\item{Sw}{White noise power (m^2/Hz).}
}
\value{
The SHOW PSD evaluated at \code{fseq} (m^2/Hz).
}
\description{
Calculate SHOW PSD.
}
