% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showf_fit_nls.R
\name{showf_fit_nls}
\alias{showf_fit_nls}
\title{SHOWF fit method non-linear least squares (NLS)}
\usage{
showf_fit_nls(
  fseq,
  Ypsd,
  fs,
  Temp,
  bin_size,
  bin_type,
  phi0,
  fit_type = c("direct", "incremental"),
  optimizer = c("optim", "Adam"),
  getHessian = FALSE,
  getJacobian = FALSE,
  ...
)
}
\arguments{
\item{fseq}{Frequency vector (units: Hz).}

\item{Ypsd}{PSD vector.}

\item{fs}{Sampling frequency.}

\item{Temp}{True temperature.}

\item{bin_size}{Bin size.}

\item{bin_type}{Either "mean" or "median".  The former is more efficient, the latter is more robust, so better for preliminary SHOW fit to denoise against.}

\item{phi0}{Parameter value (transformed scale) to initialize optimization.}

\item{fit_type}{If "direct", fit all parameters at once.  If "incremental", do one, then two, then three, etc.}

\item{optimizer}{Either "optim" (in R) or "Adam" (supplied by realPSD). For now, Adam's hyperparameter tunning is not fully supported. Learning rate and nsteps are preset. \code{Adam} is only used for \code{direct} fitting.}

\item{getHessian}{If TRUE, return the numerical Hessian matrix of the original SHOW parameter f0, Q, Rw}

\item{...}{Additional arguments to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
A list with elements
\describe{
\item{\code{par}}{The fitted parameter value which is the minimizer of the objective function.}
\item{\code{Temp}}{The temperature constant used in calculation, Kelvin.}
\item{\code{value}}{The value of the objective function (negative loglikelihood) at the given \code{par}.}
\item{\code{he}}{The numerical Hessian matrix of the SHOW model parameter f0, Q, Rw.}
\item{\code{jac}}{The numerical Jacobian matrix of the SHOW model parameter f0, Q, Rw.}
\item{\code{exitflag}}{The exit flag given by the optimizer, for debug purposes}
}
}
\description{
SHOWF fit method non-linear least squares (NLS)
}
