% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showf_psd.R
\name{showf_psd}
\alias{showf_psd}
\title{Calculate SHOWF PSD.}
\usage{
showf_psd(fseq, k, f0, Q, Sw, Af, alpha, Temp)
}
\arguments{
\item{fseq}{Frequency vector (units: Hz).}

\item{k}{Cantilever stiffness (N/m).}

\item{f0}{Resonance frequency (Hz).}

\item{Q}{Quality factor (unitless).}

\item{Sw}{White noise power (m^2/Hz).}
}
\value{
The SHOW PSD evaluated at \code{fseq} (m^2/Hz).
}
\description{
Calculate SHOWF PSD.
}
\note{
Basically the same as psdSHO.R but psdSHO offers the choice of unit conversion from m2/Hz to fm2/Hz.
}
