% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_time.R
\name{sim_time}
\alias{sim_time}
\title{Simulate a Gaussian time-domain signal from its continuous-time PSD.}
\usage{
sim_time(fs, N, fseq, cpsd)
}
\arguments{
\item{fs}{Sampling frequency.}

\item{N}{Total number of observations.}

\item{fseq}{Vector of (ordered) frequencies at which the continuous-time PSD is provided.}

\item{cpsd}{Vector of continuous-time PSD values corresponding to \code{fseq}.}
}
\value{
Simulated time series (real numeric vector).
}
\description{
Simulate a Gaussian time-domain signal from its continuous-time PSD.
}
