% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_fft.R
\name{sin_fft}
\alias{sin_fft}
\title{FFT (discrete Fourier transform) of sine wave noise}
\usage{
sin_fft(fseq, fs, D, xi, phi)
}
\arguments{
\item{fseq}{A sequence of frequency points (must be points in the frequency basis)}

\item{fs}{Sampling frequency}

\item{D}{Amplitude}

\item{xi}{as specified in the paper}

\item{phi}{as specified in the paper}
}
\description{
FFT (discrete Fourier transform) of sine wave noise
}
