% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unwrap}
\alias{unwrap}
\title{Unwrap radian phases by adding multiples of 2*pi as appropriate to
remove jumps greater than tol. tol defaults to pi.}
\usage{
unwrap(a, tol = pi, dim = 1)
}
\arguments{
\item{a}{Radian angle.}

\item{tol}{Tolerance.}

\item{dim}{Dimension.}
}
\description{
Unwrap radian phases by adding multiples of 2*pi as appropriate to
remove jumps greater than tol. tol defaults to pi.
}
