% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Calculates the one-sided periodogram from given time series and sampling frequency SF
require(fftw)}
\usage{
periodogram(yTime, SF_s, T_s)
}
\arguments{
\item{yTime}{Time series to transform}

\item{SF_s}{Sampling frequecy (Hz)}

\item{T_s}{Total time (seconds)}
}
\value{
List of xFreq: One-sided periodogram frequencies and yFreq: One-sidede periodogram ordinates
}
\description{
Calculates the one-sided periodogram from given time series and sampling frequency SF
require(fftw)
}
