% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_denoise.R
\name{psd_denoise}
\alias{psd_denoise}
\title{PSD denoise}
\usage{
psd_denoise(fseq, psd_noise, Q, f0, tau, Rw, freq_range)
}
\arguments{
\item{fseq}{Frequency vector}

\item{psd_noise}{Noisy PSD}

\item{Q}{Quality factor}

\item{f0}{Resonance frequency}

\item{tau}{sig2}

\item{Rw}{Aw/sig2}

\item{freq_range}{frequency range, i.e. f0 +- f0/sqrt(2)}
}
\description{
PSD denoise
}
