% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_denoise.R
\name{psd_denoise}
\alias{psd_denoise}
\title{PSD denoise}
\usage{
psd_denoise(fseq, psd_noise, Q, f0, k, Temp, unit_conversion, Aw,
  freq_range)
}
\arguments{
\item{fseq}{Frequency vector}

\item{psd_noise}{Noisy PSD}

\item{Q}{Quality factor}

\item{f0}{Resonance frequency}

\item{k}{k}

\item{Temp}{Temperature}

\item{unit_conversion}{indicator if unit conversion is required}

\item{Aw}{White noise amplitutde}

\item{freq_range}{frequency range, i.e. f0 +- f0/sqrt(2)}
}
\description{
PSD denoise
}
