% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_denoise.R
\name{psd_denoise}
\alias{psd_denoise}
\title{Remove periodic noise from the periodogram.}
\usage{
psd_denoise(Yf, psd_fs, alpha = 0.01)
}
\arguments{
\item{Yf}{Vector of periodogram ordinates.}

\item{psd_fs}{Discrete-time PSD at the values in \code{Yf}, i.e., theoretical expectation of \code{Yf} in the absence of noise.}

\item{alpha}{Type-I error tolerance level.}
}
\value{
The vector \code{Yf} with outliers replaced by random Exponential draws with mean given by \code{psd_d}.
}
\description{
Remove periodic noise from the periodogram.
}
