% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psdSHO.R
\name{psdSHO}
\alias{psdSHO}
\title{Power spectral density of SHO model.}
\usage{
psdSHO(f, f0, Q, k, Kb, T, unit_conversion = TRUE)
}
\arguments{
\item{f}{Frequency or a vector of frequencies in Hz.}

\item{Q}{Quality factor.}

\item{k}{Cantilever stiffness.}

\item{Kb}{Boltzmann constant.}

\item{T}{Temperature in Kelvin.}

\item{unit_conversion}{Indicator of whether or not to convert the unit of PSD to fm2/Hz. The femtometer (fm) is an SI unit of length equal to 1e-15 meters.}

\item{f_0}{Resonance frequency f_0.}
}
\description{
Power spectral density of SHO model.
}
