% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showf_psd.R
\name{showf_psd}
\alias{showf_psd}
\title{Calculate SHOWF PSD.}
\usage{
showf_psd(freq, k, f0, Q, Sw, Sf, alpha, Temp)
}
\arguments{
\item{freq}{Frequency vector (units: Hz).}

\item{k}{Cantilever stiffness (N/m).}

\item{f0}{Resonance frequency (Hz).}

\item{Q}{Quality factor (unitless).}

\item{Sw}{Power of the white noise term (m^2/Hz).}

\item{Sf}{Power of the 1/f noise term (m^2/Hz).}

\item{alpha}{Exponent of the 1/f noise term (unitless).}

\item{Temp}{Temperature (Kelvin).}
}
\value{
The SHOWF PSD evaluated at \code{freq} (m^2/Hz).
}
\description{
Calculate SHOWF PSD.
}
