% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_fft.R
\name{sin_fft}
\alias{sin_fft}
\title{Discrete Fourier transform of a sine wave.}
\usage{
sin_fft(freq, fs, N, A, xi, phi)
}
\arguments{
\item{freq}{Frequency vector.}

\item{fs}{Sampling frequency.}

\item{N}{Number of observations (see 'Details').}

\item{A}{Amplitude of the sine wave (see 'Details').}

\item{xi}{Frequency (i.e., inverse-period) of the sine wave (see 'Details').}

\item{phi}{Phase of the sine wave (see 'Details').}
}
\value{
A vector of the same length at \code{freq} containing the continuously-extended discrete Fourier transform of the sine wave evaluated at time points \code{0:(N-1)/fs}.
}
\description{
Discrete Fourier transform of a sine wave.
}
\details{
For a sine wave defined as\preformatted{g(t) = A * sin(2*pi*xi * t + phi),
}

computes the (continuously-extended) discrete Fourier transform\preformatted{G(f) = sum_\{n=0\}^\{N-1\} g(n/fs) * exp(2*pi*i/fs * f)
}

at the frequencies in \code{freq}.  When \code{freq = 0:(N-1) * fs/N}, \code{G(freq)} is the FFT of \code{g(0:(N-1)/fs)}.
}
